api/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==2.1.3
Django==3.1.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-extra-fields==3.0.2
django-filter==2.4.0
django-simple-history==2.12.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-extensions==0.7.0
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
